[FRAMEDYN] Partplement framedyn.dll So far, only CHString class is implemented in it
Regarding CHString class implementation:
It has been implemented matching Windows behavior and bugs
It might not be really nice at some points but this ensures
compatibility with applications relying on it.
As a side note: this might explain why this API was abandonned after w2k3...

Regarding the DLL itself:
Its implementation into ReactOS allows systeminfo.exe application from w2k3 to
load and start initialization into ReactOS. It stops working later on with some
IDispatch error.
The DLL cannot be used to build any application against framedyn.dll with our "default" BE (read with G++).
Only MSVC mangled name are exported so far (as on Windows ;-)) thanks to aliasing. This is why there are two spec files.
This makes G++ unable to link with it though. This is the reason why there are no framedyn_apitests so far.
Furthermore, the DLL only have exports for x86 and none for x64 for the moment.

Any help is welcome for having exports for x64 target and to be able to have aliasing for G++ (from its mangled names
to MSVC ones).

svn path=/trunk/; revision=59940