[PSAPI_APITEST] Add an apitest for psapi GetDeviceDriverFileName.
I'm looking for help to bring it even farther. Ideally, it would be interesting
to be able to GetDeviceDriverFileName on ntoskrnl base address. The whole point is
about getting it dynamically.

The day we can do it properly, I can predict that it will fail on ReactOS, we're not having
correct paths for KDCOM, HAL, and NTOSKRNL modules in the kernel (thank you FreeLdr? - Where are you
starting '\'?)

svn path=/trunk/; revision=60566