[EXT2LIB] Whatever happened during format, always unmount the volume. This allows switching FS in usetup after formating an ext2 volume.
As a record, before (read, with the IopParseDevice) hack, this wasn't possible and if formating with ext2 was possible, kernel would have written with either RawFS or FastFAT driver. Which is not exactly what we want ;-).
But no joy guys... The Ext2Fsd needs to work in ReactOS first before we can effectively install ReactOS on ext2 volumes.

In any case, that's a clear step forward :-)

svn path=/trunk/; revision=65174