[KMTESTS:IO] Extend the IoCreateFile tests with two parts tests. First part properly checks for opening/reparse handling on mount points (and ideally on symlinks, see below). The other part checks for proper handling of the IO_STOP_ON_SYMLINK parameter.
Regarding symlink, I'm clearly not convinced that what I did is correct (hence the failure on both W2K3 & ROS). It seems to me that symlink resolution is up to the FSD and not to IO which only handles mount point traversal.
If someone (Alex?) can confirm/infirm and give more insight, that's more than welcome.

svn path=/trunk/; revision=69322