[COMCTL32] comctl32/propsheet: Fix redrawing of the static text elements.
Partial revert of SVN r74136 (b1b4279): enforce the correct "color" for
the background of the static text elements: indeed, using a "transparent"
background kept instead the old window color, which was not correctly
repainted, unless you enforced a repaint by, e.g. moving away then back
the window, or moved something else on top of it.

Tested and approved by 'reactosfanboy'.
Was detected as a regression in the wizard of DVDWrite Now 1.5.12
Formatter GUI.

CORE-13929
CORE-12912